+++ src/main/java/com/example/messages/AbortImposeResponse.java
@@ -1,13 +1,13 @@
 package com.example.messages;
 
 public class AbortImposeResponse {
-    int ballot;
+    long ballot;
 
-    public AbortImposeResponse(int _ballot) {
+    public AbortImposeResponse(long _ballot) {
         ballot = _ballot;
     }
 
-    public int getBallot() { // needed to distinguish abort from other older requests
+    public long getBallot() { // needed to distinguish abort from other older requests
         return ballot;
     }
 }
